[Parley] refactor: Decompose ScrapManager.cs from 1,589 to 672 lines (#1271)
Extracted ScrapSerializationService (serialize/deserialize/load/save/migrate),
ScrapRestoreService (restore operations with shared validation), and moved
ScrapEntry/ScrapData/RestoreResult to Models/. All 671 tests pass.
